# applications/services/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for school project",
        "title": "Book \u0026 Venyl Loan Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/vinyl/all/": {
            "get": {
                "description": "Gets a list of all vinyls",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Vinyl"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/get-by-title/{title}": {
            "get": {
                "description": "Gets a vinyl by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of vinyl",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Vinyl"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/get-recs-artist/{title}": {
            "get": {
                "description": "Gets a recommended list of all vinyls from artist by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of vinyl",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.VinylSimple"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/get-recs-year/{title}": {
            "get": {
                "description": "Gets a recommended list of all vinyls from year by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of vinyl",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.VinylSimple"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/get-simple/{title}": {
            "get": {
                "description": "Gets a simplified list of vinyls by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of vinyl",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.VinylSimple"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/get/{id}": {
            "get": {
                "description": "Gets a vinyl by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of vinyl",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Vinyl"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/vinyl/search/{title}": {
            "get": {
                "description": "Gets a list of vinyls from search by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vinyl"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of vinyl",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Vinyl"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.Vinyl": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "artist": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "main.VinylSimple": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    }
}